Use the layerViews instead of the placeLayers to determine map bounds.
A mapView's placeLayers may not represent all the places on a map; oly the ones that are visible in the initial viewport.  A layerView has no layer if it's out of bounds.  This was causing a bug where placeBounds was undefined if all the places fell out of the map viewport bounds.